Added 2 States and Keyboard Clickability
Added 2 states. One set to a matrix of empty strings that will update as we add letters. The other state keeps track of the position on the matrix where we will add our next letter. Also made it so we can click the keyboard and add letters to the first row of boxes.